requirements.txt
playwright
customtkinter
python-dotenv
google-generativeai